refactor: move transaction function to dynamoDb util
