+++ server/services/productService.js
@@ -5,7 +5,7 @@ const {
 const {
   generateTransaction,
   transactWrite,
-} = require("../utils/dynamoDbTransaction");
+} = require("../utils/dynamoDb");
 
 const PRODUCTS_TABLE_NAME = process.env.PRODUCTS_TABLE;
 const STOCKS_TABLE_NAME = process.env.STOCKS_TABLE;
