+++ server/utils/dynamoDb.js
@@ -3,6 +3,7 @@ const {
   DynamoDBDocumentClient,
   GetCommand,
   ScanCommand,
+  TransactWriteCommand,
 } = require("@aws-sdk/lib-dynamodb");
 
 const client = new DynamoDBClient({
@@ -39,4 +40,45 @@ const getItemFromDynamoDb = async (tableName, key) => {
   }
 };
 
-module.exports = { dynamoDb, getAllItemsFromDynamoDb, getItemFromDynamoDb };
+const generateTransaction = (operations) => {
+  const transactItems = operations.map((operation) => {
+    const { Action, TableName, Item } = operation;
+
+    if (!TableName || !Action || !Item) {
+      throw new Error(
+        "Each operation must have TableName, Action, and either Item or Key"
+      );
+    }
+
+    const transactionItem = {
+      [Action]: {
+        TableName,
+        Item,
+      },
+    };
+
+    return transactionItem;
+  });
+
+  return { TransactItems: transactItems };
+};
+
+const transactWrite = async (params) => {
+  try {
+    const command = new TransactWriteCommand(params);
+    const response = await dynamoDb.send(command);
+
+    return response;
+  } catch (error) {
+    console.error("Transaction error:", error.message);
+    throw new Error("Failed to execute transaction");
+  }
+};
+
+module.exports = {
+  dynamoDb,
+  getAllItemsFromDynamoDb,
+  getItemFromDynamoDb,
+  generateTransaction,
+  transactWrite,
+};
